+++ WEB/Src/PerformanceCollector/PerformanceCollector/Implementation/QuickPulse/QuickPulseConstants.cs
@@ -36,6 +36,11 @@
         internal const string XMsQpsMachineNameHeaderName = "x-ms-qps-machine-name";
 
         /// <summary>
+        /// Role name header.
+        /// </summary>
+        internal const string XMsQpsRoleNameHeaderName = "x-ms-qps-role-name";
+
+        /// <summary>
         /// Invariant version header.
         /// </summary>
         internal const string XMsQpsInvariantVersionHeaderName = "x-ms-qps-invariant-version";
@@ -44,6 +49,18 @@
         /// Authentication API key.
         /// </summary>
         internal const string XMsQpsAuthApiKeyHeaderName = "x-ms-qps-auth-api-key";
+
+        /// <summary>
+        /// Service polling interval hint.
+        /// </summary>
+        /// <remarks>Contains a recommended time (in milliseconds) before we ping the service again.</remarks>
+        internal const string XMsQpsServicePollingIntervalHintHeaderName = "x-ms-qps-service-polling-interval-hint";
+
+        /// <summary>
+        /// Service endpoint redirect.
+        /// </summary>
+        /// <remarks>Contains a URI of the service endpoint we must permanently use <b>for the particular resource</b>.</remarks>
+        internal const string XMsQpsServiceEndpointRedirectHeaderName = "x-ms-qps-service-endpoint-redirect";
 
         /// <summary>
         /// The following authentication headers must be received and submitted back to the service with no modification.