+++ WEB/Src/PerformanceCollector/PerformanceCollector/QuickPulseTelemetryModule.cs
@@ -12,6 +12,7 @@
     using Microsoft.ApplicationInsights.DataContracts;
     using Microsoft.ApplicationInsights.Extensibility;
     using Microsoft.ApplicationInsights.Extensibility.Filtering;
+    using Microsoft.ApplicationInsights.Extensibility.Implementation;
     using Microsoft.ApplicationInsights.Extensibility.Implementation.Tracing;
     using Microsoft.ApplicationInsights.Extensibility.PerfCounterCollector.Implementation;
     using Microsoft.ApplicationInsights.Extensibility.PerfCounterCollector.Implementation.QuickPulse;
@@ -201,7 +202,8 @@
                             this.OnStopCollection,
                             this.OnSubmitSamples,
                             this.OnReturnFailedSamples,
-                            this.OnUpdatedConfiguration);
+                            this.OnUpdatedConfiguration,
+                            this.OnUpdatedServiceEndpoint);
 
                         this.CreateStateThread();
 
@@ -238,7 +240,7 @@
 
                     if (this.ServiceClient != null)
                     {
-                        quickPulseTelemetryProcessor.ServiceEndpoint = this.ServiceClient.ServiceUri;
+                        quickPulseTelemetryProcessor.ServiceEndpoint = this.ServiceClient.CurrentServiceUri;
                     }
 
                     QuickPulseEventSource.Log.ProcessorRegistered(this.TelemetryProcessors.Count.ToString(CultureInfo.InvariantCulture));
@@ -260,7 +262,7 @@
             }            
         }
 
-        private static string GetInstanceName(TelemetryConfiguration configuration)
+        private static CloudContext GetCloudContext(TelemetryConfiguration configuration)
         {
             // we need to initialize an item to get instance information
             var fakeItem = new EventTelemetry();
@@ -274,7 +276,7 @@
                 // we don't care what happened there
             }
 
-            return string.IsNullOrWhiteSpace(fakeItem.Context?.Cloud?.RoleInstance) ? Environment.MachineName : fakeItem.Context.Cloud.RoleInstance;
+            return fakeItem.Context?.Cloud;
         }
 
         private static string GetStreamId()
@@ -403,7 +405,9 @@
             }
 
             // create the default production implementation of the service client with the best service endpoint we could get
-            string instanceName = GetInstanceName(configuration);
+            CloudContext cloudContext = GetCloudContext(configuration);
+            string instanceName = string.IsNullOrWhiteSpace(cloudContext?.RoleInstance) ? Environment.MachineName : cloudContext.RoleInstance;
+            string roleName = cloudContext?.RoleName ?? string.Empty;
             string streamId = GetStreamId();
             var assemblyVersion = SdkVersionUtils.GetSdkVersion(null);
             bool isWebApp = PerformanceCounterUtility.IsWebAppRunningInAzure();
@@ -411,6 +415,7 @@
             this.ServiceClient = new QuickPulseServiceClient(
                 serviceEndpointUri,
                 instanceName,
+                roleName,
                 streamId,
                 this.ServerId,
                 assemblyVersion,
@@ -618,7 +623,7 @@
                 {
                     telemetryProcessor.StartCollection(
                         this.dataAccumulatorManager,
-                        this.ServiceClient.ServiceUri,
+                        this.ServiceClient.CurrentServiceUri,
                         this.config,
                         this.DisableFullTelemetryItems);
                 }
@@ -707,7 +712,19 @@
             return errorsConfig.Concat(errorsPerformanceCounters).ToArray();
         }
 
-#endregion
+        private void OnUpdatedServiceEndpoint(Uri newServiceEndpoint)
+        {
+            QuickPulseEventSource.Log.TroubleshootingMessageEvent("Service endpoint updated.");
+            
+            lock (this.telemetryProcessorsLock)
+            {
+                foreach (var telemetryProcessor in this.TelemetryProcessors)
+                {
+                    telemetryProcessor.ServiceEndpoint = newServiceEndpoint;
+                }
+            }
+        }
+        #endregion
 
         /// <summary>
         /// Dispose implementation.