+++ examples/WorkerServiceSDK/BackgroundTasksWithHostedService/BackgroundTasksWithHostedService/Program.cs
@@ -1,0 +1,64 @@
+﻿using System.Threading.Tasks;
+using BackgroundTasksWithHostedService.HostedServices;
+using Microsoft.ApplicationInsights.Extensibility;
+using Microsoft.ApplicationInsights.Extensibility.PerfCounterCollector.QuickPulse;
+using Microsoft.Extensions.Configuration;
+using Microsoft.Extensions.DependencyInjection;
+using Microsoft.Extensions.Hosting;
+using Microsoft.Extensions.Logging;
+
+namespace BackgroundTasksWithHostedService
+{
+    public class Program
+    {
+        public static async Task Main(string[] args)
+        {
+            var host = new HostBuilder()
+                .ConfigureLogging((hostContext, config) =>
+                {
+                    config.AddConsole();
+                    config.AddDebug();
+                })
+                .ConfigureHostConfiguration(config =>
+                {
+                    config.AddEnvironmentVariables();
+                })
+                .ConfigureAppConfiguration((hostContext, config) =>
+                {
+                    config.AddJsonFile("appsettings.json", optional: true);
+                    config.AddJsonFile($"appsettings.{hostContext.HostingEnvironment.EnvironmentName}.json", optional: true);
+                    config.AddCommandLine(args);
+                })
+                .ConfigureServices((hostContext, services) =>
+                {
+                    services.AddLogging();
+                    services.AddHostedService<TimedHostedService>();
+
+                    // Application Insights
+                    // Add custom TelemetryInitializer
+                    services.AddSingleton<ITelemetryInitializer, MyCustomTelemetryInitializer>();
+
+                    // Add custom TelemetryProcessor
+                    services.AddApplicationInsightsTelemetryProcessor<MyCustomTelemetryProcessor>();
+
+                    // Example on Configuring TelemetryModules.
+                    // [SuppressMessage("Microsoft.Security", "CS002:SecretInNextLine", Justification="Not a real api key, this is example code.")]
+                    services.ConfigureTelemetryModule<QuickPulseTelemetryModule>((mod, opt) => mod.AuthenticationApiKey = "put_actual_authentication_key_here");
+
+                    // instrumentation key is read automatically from appsettings.json
+                    services.AddApplicationInsightsTelemetryWorkerService();
+                })
+                .UseConsoleLifetime()
+                .Build();
+
+            using (host)
+            {
+                // Start the host
+                await host.StartAsync();
+
+                // Wait for the host to shutdown
+                await host.WaitForShutdownAsync();
+            }
+        }
+    }
+}